feat(kernel): set up and load IDT
This commit defines the IDT entry for the divide by 0 exception and
loads the IDT using the lidt instruction. It also implements a handler
function that prints an error message and halts the CPU when the
exception occurs. To test the functionality, a division by 0 is
performed in the main code. The expected output is 'Divide by 0 error'
(in fact character 'D' is displayed) in red color on the screen.