feat(kernel): add push/pop macros
This commit defines two macros: push_regs and pop_regs, which push and
pop all the general-purpose registers and the rflags register to and
from the stack, respectively. These macros are used in the interrupt
handler routines to save and restore the register state before and after
handling the interrupt. This makes the code more concise, readable, and
maintainable.