feat(kernel): jump to UserEntry in ring 3
This commit adds the code for switching from ring 0 to ring 3 by
pushing the values for CS, RFLAGS, offset, interrupt number, DS, and
UserEntry address on the stack and returning from interrupt. It also
jumps to the user entry point in ring 3 and writes 'U' in white color
to the video memory. This is part of the kernel code for a 64-bit
operating system.